+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
 * Adding Material Accent and Foreground colors to Configuration.h.
 * Remove extra space between slide menu and top menu.
+* Force Active focus on TextField in sending message after each send.
 
 
 ## [0.0.21]